refactor(email): move the BYO readiness invalidation into the email module
The generic credential store carried email-specific logic; the email module
that defines the BYO provider list now owns the invalidation and exposes a
type-predicate membership check, removing a readonly-array cast.

Part of LUM-3501
